+++ Chimera.js
@@ -1,6 +1,7 @@
 import {ChimeraApplication} from "./Chimera/Application.js";
 import {Configuration} from "./Chimera/Configuration.js";
 import {DarwinWindowController} from "./Chimera/DarwinWindowController.js";
+import {LinuxWindowController} from "./Chimera/LinuxWindowController.js";
 import {SessionController} from "./Chimera/SessionController.js";
 import {SurfaceController} from "./Chimera/SurfaceController.js";
 import {WindowController} from "./Chimera/WindowController.js";
@@ -8,6 +9,7 @@ import {WindowController} from "./Chimera/WindowController.js";
 export {ChimeraApplication} from "./Chimera/Application.js";
 export {Configuration} from "./Chimera/Configuration.js";
 export {DarwinWindowController} from "./Chimera/DarwinWindowController.js";
+export {LinuxWindowController} from "./Chimera/LinuxWindowController.js";
 export {SessionController} from "./Chimera/SessionController.js";
 export {SurfaceController} from "./Chimera/SurfaceController.js";
 export {WindowController} from "./Chimera/WindowController.js";
@@ -19,6 +21,7 @@ export const Chimera = {
 	ChimeraApplication,
 	Configuration,
 	DarwinWindowController,
+	LinuxWindowController,
 	WindowController,
 	SessionController,
 	SurfaceController,